refactor: extract cli output path selection
